git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@20453 5aad4ff5-4b58-4159-ba3f-0e6328df4cad

+++ tools/disthelper/windows/icon.cpp
@@ -139,7 +139,9 @@ int CommandSetIcon(std::string aExecutable, std::string aIconFile, std::string a
   }
 
   // Allocate the group resource entry
-  if (sizeof(IconHeader) + header->ImageCount * sizeof(IconDirEntry) > filesize) {
+  if ((long)(sizeof(IconHeader) + header->ImageCount * sizeof(IconDirEntry)) >
+      filesize)
+  {
     OutputDebugString(_T("Inconsistent icon header.\n"));
     free(data);
     return DH_ERROR_PARSE;
